+++ infrastructure/src/main/java/com/kaua/monitoring/infrastructure/profile/ProfileAvatarGateway.java
@@ -88,7 +88,9 @@ public class ProfileAvatarGateway implements AvatarGateway {
 
             final var avatarAlreadyExists = s3Client.listObjects(avatarAlreadyExistsRequest);
 
-            avatarAlreadyExists.contents().forEach(image -> deleteAvatar(image.key()));
+            if (!avatarAlreadyExists.contents().isEmpty()) {
+                avatarAlreadyExists.contents().forEach(image -> deleteAvatar(image.key()));
+            }
         } catch (Throwable t) {
             throw new RuntimeException(t);
         }
